Show books on account page
I agree with Ming that it is not feasible to show all the user's
books on the History page. This code displays all the books that
were uploaded by the user on his/her account page. The layout here
could maybe be enriched by showing whether the book has been
requested yes or no.